+++ lib/src/factories/dart_ddi_base_factory.dart
@@ -123,4 +123,5 @@ abstract class DDIBaseFactory<BeanT extends Object> with InstanceFactoryMixin {
   ///
   /// - `ddiInstance`: The DDI instance to use for operations.
   Future<void> dispose({required DDI ddiInstance});
+
 }